Direct output to /dev/null.
Couscous doesn't allow force pushing and fails if documentation hasn't changed.
This redirects the output of the command to /dev/null so that we don't block CI.

We should add a custom script to inspect the output and correctly validate that the deploy command is working correctly
even if the remote hasn't changed.